feat: Construção da tela de cadastro
